Rename test server file
